Moves hamburger-button inside menu
