--- components/hamburger-button/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@oma-wc/hamburger-button",
-  "version": "0.11.0",
-  "description": "",
-  "main": "dist/index.cjs.js",
-  "unpkg": "dist/index.cjs.js",
-  "module": "dist/index.es.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/oma-wc/oma.git",
-    "directory": "components/hamburger-button"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "GPL3",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "haunted": "^4.7.0"
-  }
-}
